+++ src/__tests__/hooks/outputFiles.test.ts
@@ -5,13 +5,13 @@
 import fs from 'fs-extra';
 import path from 'path';
 
-import { mockReport } from '../../__tests__/helpers/testHelpers';
+import { mockReport } from '../helpers/testHelpers';
 import { OutputOptions } from '../../types';
 
 describe('Output Files', () => {
     const directoryName = '/test/';
     const init = async (output: OutputOptions, context: string) => {
-        const { hooks } = require('../outputFiles');
+        const { hooks } = require('../../hooks/outputFiles');
         await hooks.output.call(
             // eslint-disable-next-line no-console
             { log: console.log, options: { output, context } },
@@ -47,7 +47,7 @@ describe('Output Files', () => {
         );
 
         test('It should export hooks', () => {
-            const outputFiles = require('../outputFiles');
+            const outputFiles = require('../../hooks/outputFiles');
 
             expect(typeof outputFiles.hooks).toBe('object');
         });
